Refactor asv to output tests based on continuous grasper
