+++ polymetis/README.md
@@ -36,6 +36,16 @@ All documentation on the [website](https://polymetis-docs.github.io/). Includes:
 - Writing developing your own custom controllers in PyTorch
 - Full [autogenerated documentation](https://polymetis-docs.github.io/modules.html)
 
+## Benchmarking
+
+To run benchmarking, first configure the [script](polymetis/tests/python/polymetis/benchmarks/benchmark_robustness.py) to point to your hardware instance, then run
+
+```bash
+asv run --python=python --set-commit-hash $(git rev-parse HEAD)
+```
+
+Commit the result under `.asv/results`; it will show up under the benchmarking page in the documentation.
+
 ## Citing
 If you use Polymetis in your research, please use the following BibTeX entry.
 ```
