+++ polymetis/docs/source/index.rst
@@ -33,6 +33,10 @@ Autogenerated docs
 
    modules
 
+Benchmarks
+---------------------------
+
+To see benchmarks over time (using `Airspeed Velocity <https://asv.readthedocs.io/>`_), `click here <../_static/asv-benchmarks/index.html>`_.
 
 Indices and tables
 ------------------
