+++ polymetis/polymetis/tests/python/polymetis/benchmarks/benchmark_robustness.py
@@ -1,7 +1,13 @@
-import random
+from polymetis.utils.continuous_grasper import ManipulatorSystem
 
 
-def track_num_successes_per_1000():
-    return random.randint(900, 1000)
+def track_num_successes_per_100():
+    iters = 2
+    robot_kwargs = {}
+    gripper_kwargs = {}
+    robot = ManipulatorSystem(robot_kwargs, gripper_kwargs)
+    total_successes, total_tries = robot.continuously_grasp(iters)
+    return total_successes
 
-track_num_successes_per_1000.unit = "num_successes"
+
+track_num_successes_per_100.unit = "num_successes"
